Set default pool size to 1
This also simplifies the logic to determine whether a given JDBC
statement has results, required by `Statement#execute` API.

Previously, the statement implementation would open a new session
and transaction, request and parse the execution plan of the query
that had just run to determine whether it had results, while the
original transaction was still running.

However, that approach breaks down when the pool size is set to 1,
as a new session would require a new connection that is just not
available, leading a (default) timeout of 1 minute. To make things
worse, the timeout exception is ignored and the query is assumed
to return no result, after waiting for the whole timeout duration
to elapse.

This commit now relies on the Neo4j driver's `Result#hasNext` that
will simply peek into the streaming records and see if there is at
least one. This is much more efficient since it does not require
any roundtrip. It is also more accurate, since the query execution
plan is not requested at a later time, during which the database
state could have changed.