Work around the priorization of change class selection in core
Database-specific change subclasses (like
`liquibase.ext.neo4j.change.LoadGraphDataChange` or
`liquibase.ext.neo4j.change.InsertNodeChange`) can interfere with
executions not targeting Neo4j.

The reason for this behavior is that
`liquibase.change.ChangeFactory#create` selects a particular
class for a given change, only based on the class' declared priority.

The two above changes declare a `PRIORITY_DATABASE`, which means they
have higher priority than the built-in changes.

When Neo4j is in the classpath, this means these classes can get picked
up, even though:

 - a completely different DBMS is configured at execution time
 - `liquibase.change.Change#supports` is properly implemented

If any of the overridden methods implement Neo4j-specific behavior (
Cypher queries e.g.), this leads to execution failures.

A more subtle failure that occurs is when extra fields are introduced
by the subclass (like `InsertNodeChange` and its field `labelName`) and
are initialized with non-default values.
What happens is that the checksum value changes and becomes incompatible
with what was stored in the target database.

The only viable fix is to cater for non-Neo4j databases.
When a non-Neo4j database is encountered, call the parent class
implementation instead of running the Neo4j-specific implementation.
This works as long as other extensions in the classpath do not also
subclass the same parent class.

Fixes #413