chore(build-nightly.yml): add nightly build workflow to build the extension against the latest Liquibase artifact on a schedule
A new workflow file `build-nightly.yml` has been added to the `.github/workflows` directory. This workflow is triggered on a schedule using a cron expression (`0 7 * * 1-5`) and is named "Nightly build".

The workflow consists of a single job named `nightly-build`. It uses the `liquibase/build-logic/.github/workflows/os-extension-test.yml@v0.5.2` action to build the extension. The `nightly` input is set to `true` to indicate that this is a nightly build.

The `secrets` configuration is set to `inherit`, which means that the secrets configured at the repository level will be inherited by this workflow.